+++ examples/basic_solana/src/lib.rs
@@ -7,7 +7,7 @@ use crate::state::{read_state, State};
 use candid::{CandidType, Deserialize, Principal};
 use sol_rpc_client::{ed25519::Ed25519KeyId, IcRuntime, SolRpcClient};
 use sol_rpc_types::{
-    CommitmentLevel, ConsensusStrategy, MultiRpcResult, RpcConfig, RpcSources, SolanaCluster,
+    CommitmentLevel, ConsensusStrategy, MultiRpcResult, RpcSources, SolanaCluster,
 };
 use solana_hash::Hash;
 use std::str::FromStr;
@@ -62,12 +62,9 @@ pub fn client() -> SolRpcClient<IcRuntime> {
         .with_rpc_sources(RpcSources::Default(
             read_state(|state| state.solana_network()).into(),
         ))
-        .with_rpc_config(RpcConfig {
-            response_size_estimate: None,
-            response_consensus: Some(ConsensusStrategy::Threshold {
-                total: Some(3),
-                min: 2,
-            }),
+        .with_consensus_strategy(ConsensusStrategy::Threshold {
+            min: 2,
+            total: Some(3),
         })
         .with_default_commitment_level(read_state(State::solana_commitment_level))
         .build()