+++ libs/client/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-05-27
+
+### Added
+
+- Add `getRecentPrioritizationFees` RPC method ([#92](https://github.com/dfinity/sol-rpc-canister/pull/92), [108](https://github.com/dfinity/sol-rpc-canister/pull/108))
+- Add `getSignaturesForAddress` RPC method ([#106](https://github.com/dfinity/sol-rpc-canister/pull/106))
+- Add `getSignatureStatuses` RPC method ([#96](https://github.com/dfinity/sol-rpc-canister/pull/96))
+- Add `getTokenAccountBalance` RPC method ([#90](https://github.com/dfinity/sol-rpc-canister/pull/90))
+- Add helper method to sign a transaction ([#113](https://github.com/dfinity/sol-rpc-canister/pull/113))
+- Add helper method to extract nonce value from a Solana account ([#117](https://github.com/dfinity/sol-rpc-canister/pull/117))
+
+### Changed
+
+- Use secure primitive types for `Pubkey`, `Signature` and `Hash` ([#98](https://github.com/dfinity/sol-rpc-canister/pull/98))
+- Use `RoundingError` from `sol_rpc_types` ([#105](https://github.com/dfinity/sol-rpc-canister/pull/105))
+
 ## [0.1.0] - 2025-04-29
 
 ### Added